# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AJVC Portfolio Intelligence Dashboard API",
    "description": "\n## 🚀 AJVC Portfolio Intelligence Dashboard - Backend API\n\nA comprehensive **portfolio management system** for venture capital firms to track portfolio companies,\nanalyze project execution, monitor commitments, and gain AI-powered insights from meeting notes.\n\n### Key Features\n\n* **📊 Company Management**: Track portfolio companies with investment details and status\n* **📈 Gantt Layer Analytics**: Pull data from Google Sheets and compute execution KPIs\n* **📝 AI-Powered Meeting Analysis**: Upload DOCX meeting notes for automatic parsing with Google Gemini\n* **✅ Commitment Tracking**: Automatically extract and track actionable commitments from meetings\n* **🔒 Secure Authentication**: JWT-based authentication for all protected endpoints\n\n### Tech Stack\n\n* **FastAPI** for high-performance async API\n* **PostgreSQL** with SQLAlchemy ORM\n* **Google Gemini AI** for intelligent meeting parsing\n* **Google Sheets API** for Gantt data integration\n* **JWT Authentication** for secure access\n\n### Getting Started\n\n1. Authenticate via `/api/v1/auth/login` to get a JWT token\n2. Include the token in the `Authorization: Bearer <token>` header\n3. Create or select a company via `/api/v1/companies`\n4. Upload meeting notes or pull Gantt data for analysis\n\n### API Versioning\n\nCurrent version: **v1.0.0** - All endpoints are prefixed with `/api/v1`\n    ",
...
  "paths": {
...
    "/api/v1/companies/": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "List Companies",
        "operationId": "list_companies_api_v1_companies__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyList"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "companies"
        ],
        "summary": "Create Company",
        "operationId": "create_company_api_v1_companies__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/companies/{company_id}": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Get Company",
        "operationId": "get_company_api_v1_companies__company_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "companies"
        ],
        "summary": "Update Company",
        "operationId": "update_company_api_v1_companies__company_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "companies"
        ],
        "summary": "Delete Company",
        "operationId": "delete_company_api_v1_companies__company_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/companies/{company_id}/commitments": {
      "get": {
        "tags": [
          "commitments"
        ],
        "summary": "List Commitments",
        "description": "List commitments for a company.\nSorted with overdue first, then by due date.\n\nQuery params:\n- status_filter: Filter by status (open, due-soon, overdue, resolved)",
        "operationId": "list_commitments_api_v1_companies__company_id__commitments_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitmentList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{company_id}/commitments/{commitment_id}": {
      "get": {
        "tags": [
          "commitments"
        ],
        "summary": "Get Commitment",
        "description": "Get a specific commitment.",
        "operationId": "get_commitment_api_v1_companies__company_id__commitments__commitment_id__get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          },
          {
            "name": "commitment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Commitment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitmentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "commitments"
        ],
        "summary": "Update Commitment",
        "description": "Update commitment status or details.",
        "operationId": "update_commitment_api_v1_companies__company_id__commitments__commitment_id__patch",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          },
          {
            "name": "commitment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Commitment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommitmentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommitmentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommitmentList": {
        "properties": {
          "commitments": {
            "items": {
              "$ref": "#/components/schemas/CommitmentRead"
            },
            "type": "array",
            "title": "Commitments"
          }
        },
        "type": "object",
        "required": [
          "commitments"
        ],
        "title": "CommitmentList"
      },
      "CommitmentRead": {
        "properties": {
          "person": {
            "type": "string",
            "title": "Person",
            "description": "Person responsible for the commitment",
            "examples": [
              "John Doe (CEO)",
              "Jane Smith (CTO)",
              "Founder"
            ]
          },
          "action": {
            "type": "string",
            "title": "Action",
            "description": "Description of the action to be taken",
            "examples": [
              "Complete user onboarding flow",
              "Hire 2 engineers",
              "Launch MVP"
            ]
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date",
            "description": "Due date for the commitment",
            "examples": [
              "2024-04-01"
            ]
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "Source of the commitment",
            "examples": [
              "founder-initiated",
              "aviral-pushed"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique commitment identifier"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "title": "Company Id",
            "description": "Company this commitment belongs to"
          },
          "meeting_id": {
            "type": "string",
            "format": "uuid",
            "title": "Meeting Id",
            "description": "Meeting where this commitment was mentioned"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Current status of the commitment",
            "examples": [
              "open",
              "due-soon",
              "overdue",
              "resolved"
            ]
          },
          "days_overdue": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Days Overdue",
            "description": "Number of days overdue (only for overdue commitments)",
            "examples": [
              3,
              7,
              14
            ]
          },
          "origin_meeting_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Origin Meeting Id",
            "description": "Meeting where this commitment was originally created"
          },
          "resolved_in_meeting_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved In Meeting Id",
            "description": "Meeting where this commitment was resolved"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At",
            "description": "Timestamp when commitment was marked as resolved"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Record creation timestamp"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "Last update timestamp"
          }
        },
        "type": "object",
        "required": [
          "person",
          "action",
          "source",
          "id",
          "company_id",
          "meeting_id",
          "status",
          "created_at"
        ],
        "title": "CommitmentRead"
      },
...
      "CompanyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Company name",
            "examples": [
              "TechStartup Inc",
              "FinanceAI Ltd"
            ]
          },
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompanyType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Type of company investment"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage",
            "description": "Investment stage (e.g., Seed, Series A, Series B)",
            "examples": [
              "Seed",
              "Series A",
              "Series B",
              "Growth"
            ]
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector",
            "description": "Industry sector",
            "examples": [
              "FinTech",
              "HealthTech",
              "EdTech",
              "SaaS"
            ]
          },
          "investment_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investment Date",
            "description": "Date of investment"
          },
          "investment_amount_inr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investment Amount Inr",
            "description": "Investment amount in Indian Rupees",
            "examples": [
              5000000,
              10000000
            ]
          },
          "status": {
            "$ref": "#/components/schemas/CompanyStatus",
            "description": "Current status of the company",
            "default": "Active"
          },
          "sheets_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheets Url",
            "description": "Google Sheets URL for Gantt chart data",
            "examples": [
              "https://docs.google.com/spreadsheets/d/1abc...xyz/edit"
            ]
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CompanyCreate"
      },
      "CompanyList": {
        "properties": {
          "companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyRead"
            },
            "type": "array",
            "title": "Companies"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "companies",
          "total"
        ],
        "title": "CompanyList"
      },
      "CompanyRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Company name",
            "examples": [
              "TechStartup Inc",
              "FinanceAI Ltd"
            ]
          },
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompanyType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Type of company investment"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage",
            "description": "Investment stage (e.g., Seed, Series A, Series B)",
            "examples": [
              "Seed",
              "Series A",
              "Series B",
              "Growth"
            ]
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector",
            "description": "Industry sector",
            "examples": [
              "FinTech",
              "HealthTech",
              "EdTech",
              "SaaS"
            ]
          },
          "investment_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investment Date",
            "description": "Date of investment"
          },
          "investment_amount_inr": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investment Amount Inr",
            "description": "Investment amount in Indian Rupees",
            "examples": [
              5000000,
              10000000
            ]
          },
          "status": {
            "$ref": "#/components/schemas/CompanyStatus",
            "description": "Current status of the company",
            "default": "Active"
          },
          "sheets_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheets Url",
            "description": "Google Sheets URL for Gantt chart data",
            "examples": [
              "https://docs.google.com/spreadsheets/d/1abc...xyz/edit"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "created_at"
        ],
        "title": "CompanyRead"
      },
...
      "CompanyType": {
        "type": "string",
        "enum": [
          "B2B",
          "B2C",
          "B2B2C"
        ],
        "title": "CompanyType"
      },
      "CompanyUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompanyType"
              },
              {
                "type": "null"
              }
            ]
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "investment_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investment Date"
          },
          "investment_amount_inr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investment Amount Inr"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompanyStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "sheets_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheets Url"
          }
        },
        "type": "object",
        "title": "CompanyUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}